refactor(wardrobe): split ClothingDetailScreen into focused files
Three changes in one pass:

1. Collapsed 5 copy-pasted MultiSelectSheet if-blocks into a single
   data-driven block via AttributePicker enum + nullable activePicker state
   (-83 lines of duplication)

2. Extracted ClothingAttributes, AttributeSection, AttributeChip into
   ClothingDetailComponents.kt (internal visibility). Added private
   AttributeEmptyHint helper to eliminate the repeated 5-line empty-state
   Text pattern inside ClothingAttributes.

3. Moved all @Preview functions and mock data into ClothingDetailPreviews.kt;
   removed Configuration, Preview, ClosetTheme imports from screen file.
